backend/requirements.txt
# ── Web Framework ──────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# ── AI / Extraction ────────────────────────────────────────
anthropic>=0.28.0

# ── OCR ────────────────────────────────────────────────────
pymupdf>=1.24.0
pytesseract>=0.3.10
Pillow>=10.3.0

# ── Database ───────────────────────────────────────────────
sqlalchemy>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.0

# ── Background Jobs ────────────────────────────────────────
celery[redis]>=5.4.0
redis>=5.0.0

# ── File Storage ───────────────────────────────────────────
boto3>=1.34.0

# ── HTTP Client (HubSpot) ──────────────────────────────────
httpx>=0.27.0

# ── Auth ───────────────────────────────────────────────────
svix>=1.20.0
python-jose[cryptography]>=3.3.0

# ── Config / Validation ────────────────────────────────────
pydantic>=2.7.0
pydantic-settings>=2.3.0

# ── Export ─────────────────────────────────────────────────
openpyxl>=3.1.2
